requirements.txt
# Core packages
python-dotenv==1.0.0         # For environment variable management
pandas==2.0.3                # Data manipulation and analysis
numpy==1.25.0                # Numerical operations

# Data fetching
requests==2.31.0             # HTTP requests for APIs
newsapi-python==0.2.7        # NewsAPI integration

# Database management
sqlite-utils==3.31.0         # Simplified SQLite management
sqlalchemy==2.0.20           # Database ORM
firebase-admin==6.1.0        # Firebase for cloud database (optional)

# Sentiment analysis and NLP
transformers==4.33.3         # Hugging Face models
torch==2.0.1                 # PyTorch for model support
finbert-embedding==0.1.5    # FinBERT for sentiment analysis
scikit-learn==1.3.1          # Machine learning utilities

# Summarization
openai==0.27.10              # OpenAI API for GPT models (e.g., GPT-4)
sentencepiece==0.1.99        # Tokenization for summarization models
nltk==3.8.1                  # Natural language processing (optional)

# Web application and UI
streamlit==1.25.0            # Frontend for user interaction
streamlit-aggrid==0.3.4      # Advanced table/grid integration for Streamlit

# Caching and optimization
joblib==1.3.2                # Caching and parallel processing
diskcache==5.6.1             # Disk-based caching

# Miscellaneous
dotenv-linter==0.5.0         # Linter for .env files
matplotlib==3.8.0            # Visualization for insights (optional)
seaborn==0.13.0              # Statistical data visualization
